gale2: Use #pragma once instead of inclusion guards
By now, #pragma once is supported by most compilers. There is a chance that
it is faster than inclusion guards because the compiler does not have to
check whether the macro has been undefined. Finally, there is no macro
anymore which needs to be adjusted to the file name in case of renaming.